rename: #113 DashboardAttendee -> DashboardAttendeePage로 파일명 수정
