+++ src/pages/DashboardPage/index.js
@@ -1,4 +1,4 @@
 export { default as DashboardPage } from './DashboardPage';
 export { default as DashboardInfo } from './DashboardInfo';
 export { default as DashboardEmail } from './DashboardEmail';
-export { default as DashboardAttendee } from './DashboardAttendee';
+export { default as DashboardAttendeePage } from './DashboardAttendeePage';
